requirements.txt
filelock # like fcntl but has context manager
pebble # like multiprocessing but has timeout
pyrosetta-help